requirements.txt
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-phonenumber-field==4.0.0
djangorestframework==3.10.3
idna==2.8
kavenegar==1.1.2
phonenumbers==8.11.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
